+++ applications/PfemFluidDynamicsApplication/python_scripts/fluid_mesher.py
@@ -81,88 +81,6 @@ class FluidMesher(mesher.Mesher):
         self.MeshingParameters.SetTessellationFlags(mesher_flags)
         self.MeshingParameters.SetTessellationInfo(mesher_info)
 
-
-    #
-    def SetPreMeshingProcesses(self):
-
-        if(self.echo_level>0):
-            print("::[fluid_mesher]:: -START SetPreMeshingProcesses-")
-
-        refining_parameters = self.MeshingParameters.GetRefiningParameters()
-        refining_options = refining_parameters.GetRefiningOptions()
-
-        #recover_volume_losses  = KratosPfemFluid.RecoverVolumeLosses(self.model_part, self.MeshingParameters, self.echo_level)
-        #self.mesher.SetPreMeshingProcess(recover_volume_losses)
-
-        unactive_peak_elements = False
-        unactive_sliver_elements = False
-        set_active_flag = KratosPfemFluid.SetActiveFlagMesherProcess(self.main_model_part,unactive_peak_elements,unactive_sliver_elements,self.echo_level)
-        self.mesher.SetPreMeshingProcess(set_active_flag)
-
-        inlet_management = KratosPfemFluid.InletManagement(self.model_part, self.MeshingParameters, self.echo_level)
-        self.mesher.SetPreMeshingProcess(inlet_management)
-
-        remove_mesh_nodes = KratosPfemFluid.RemoveMeshNodesForFluids(self.model_part, self.MeshingParameters, self.echo_level)
-        self.mesher.SetPreMeshingProcess(remove_mesh_nodes)
-
-        if( refining_options.Is(KratosDelaunay.MesherUtilities.REFINE_INSERT_NODES) ):
-            generate_new_nodes  = KratosPfemFluid.GenerateNewNodesBeforeMeshing(self.model_part, self.MeshingParameters, self.echo_level)
-            self.mesher.SetPreMeshingProcess(generate_new_nodes)
-
-
-    #
-    def SetPostMeshingProcesses(self):
-
-        # The order set is the order of execution:
-        if(self.echo_level>0):
-            print("::[fluid_mesher]:: -START SetPostMeshingProcesses-")
-
-
-        refining_parameters = self.MeshingParameters.GetRefiningParameters()
-        refining_options = refining_parameters.GetRefiningOptions()
-
-
-        #select mesh elements
-        select_mesh_elements  = KratosPfemFluid.SelectMeshElementsForFluids(self.model_part, self.MeshingParameters, self.echo_level)
-        self.mesher.SetPostMeshingProcess(select_mesh_elements)
-        #rebuild elements
-        #rebuild_mesh_elements = KratosDelaunay.BuildMeshElements(self.main_model_part, self.MeshingParameters, self.echo_level)
-        #self.mesher.SetPostMeshingProcess(rebuild_mesh_elements)
-
-        if( refining_options.Is(KratosDelaunay.MesherUtilities.REFINE_ADD_NODES) ):
-            select_refine_elements = KratosDelaunay.RefineElementsOnSize(self.model_part, self.MeshingParameters, self.echo_level)
-            self.mesher.SetPostMeshingProcess(select_refine_elements)
-
-        #rebuild elements
-        rebuild_mesh_elements = KratosDelaunay.GenerateNewElements(self.model_part, self.MeshingParameters, self.echo_level)
-        self.mesher.SetPostMeshingProcess(rebuild_mesh_elements)
-
-        ### rebuild boundary
-        ############ choose just one of the following two options: ############
-        ## use this if you want conditions
-        ## ATTENTION: this is slow, and must be used together with ModelMeshingWithConditionsForFluids and BuildModelPartBoundary
-        #rebuild_mesh_boundary = KratosPfemFluid.GenerateNewConditionsForFluids(self.model_part, self.MeshingParameters, self.echo_level)
-        ## if you use the following, you will not use/build/compute conditions
-        ## ATTENTION: it must be used together with ModelMeshingForFluids and BuildModelPartBoundaryForFluids
-        rebuild_mesh_boundary = KratosPfemFluid.BuildMeshBoundaryForFluids(self.model_part, self.MeshingParameters, self.echo_level)
-        #######################################################################
-        self.mesher.SetPostMeshingProcess(rebuild_mesh_boundary)
-
-
-        #unactive_peak_elements = False
-        #unactive_sliver_elements = False
-        #set_active_flag = KratosPfemFluid.SetActiveFlagMesherProcess(self.main_model_part,unactive_peak_elements,unactive_sliver_elements,self.echo_level)
-        #self.mesher.SetPostMeshingProcess(set_active_flag)
-
-
-        #inlet_management = KratosPfemFluid.InletManagement(self.model_part, self.MeshingParameters, self.echo_level)
-        #self.mesher.SetPostMeshingProcess(inlet_management)
-
-        #if( refining_options.Is(KratosDelaunay.MesherUtilities.REFINE_INSERT_NODES) ):
-            #select_refine_elements = KratosPfemFluid.SetElementsToRefineOnSize(self.model_part, self.MeshingParameters, self.echo_level)
-            #self.mesher.SetPostMeshingProcess(select_refine_elements)
-
-
     #
     def FinalizeMeshing(self):
         pass
