+++ applications/PfemFluidDynamicsApplication/python_scripts/fluid_meshing_strategy.py
@@ -18,12 +18,12 @@ class FluidMeshingStrategy(meshing_strategy.MeshingStrategy):
 
         meshers_list = []
         if( self.settings["remesh"].GetBool() and self.settings["refine"].GetBool() ):
-            meshers_list.append("KratosMultiphysics.PfemFluidDynamicsApplication.fluid_mesher")
+            meshers_list.append("KratosMultiphysics.PfemFluidDynamicsApplication.pfem_fluid_complete_mesher")
             #mesher_list.append("fluid_post_refining_mesher")
+        elif( self.settings["remesh"].GetBool() and self.settings["transfer"].GetBool() ):
+            meshers_list.append("KratosMultiphysics.PfemFluidDynamicsApplication.pfem_fluid_none_mesher")
         elif( self.settings["remesh"].GetBool() ):
-            meshers_list.append("KratosMultiphysics.DelaunayMeshingApplication.reconnect_mesher")
-        elif( self.settings["transfer"].GetBool() ):
-            meshers_list.append("KratosMultiphysics.DelaunayMeshingApplication.transfer_mesher")
+            meshers_list.append("KratosMultiphysics.PfemFluidDynamicsApplication.pfem_fluid_keeping_nodes_mesher")
 
         for mesher in meshers_list:
             full_module_name = mesher
